Fix some more compile errors.
svn path=/trunk/KDE/kdeaccessibility/kttsd/; revision=443385

+++ libkttsd/pluginproc.cpp
@@ -279,7 +279,7 @@ QString PlugInProc::getSsmlXsltFilename()
             codecName = "Unicode";
             break;
         default:
-            if ((uint)codecNum < codecList.count())
+            if (codecNum < codecList.count())
                 codecName = codecList[codecNum];
             else
             {
